Updated submodules; updated premake file Now RealEngine on build has zero warnings!

+++ RealEngine/src/RealEngine/ImGUI/ImGUILayer.cpp
@@ -61,7 +61,8 @@ namespace RealEngine {
 	void ImGUILayer::end() {
 		auto& io = ImGui::GetIO();
 		Application& app = Application::getApplication();
-		io.DisplaySize = ImVec2(app.getWindow().getWidth(), app.getWindow().getHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(app.getWindow().getWidth())
+			, static_cast<float>(app.getWindow().getHeight()));
 
 		//Render
 		ImGui::Render();
